fix __DIR__ in import type
