+++ packages/config-transformer/packages/format-switcher/src/PhpParser/Printer/FluentMethodCallPrinter.php
@@ -24,7 +24,11 @@ final class FluentMethodCallPrinter extends Standard
         $printedContent = Strings::replace($printedContent, '#^[ ]+\n#m', "\n");
 
         // remove space before " :" in main closure
-        $printedContent = Strings::replace($printedContent, '#containerConfigurator\) : void#', 'containerConfigurator): void');
+        $printedContent = Strings::replace(
+            $printedContent,
+            '#containerConfigurator\) : void#',
+            'containerConfigurator): void'
+        );
 
         return $printedContent . self::EOL_CHAR;
     }
