+++ packages/config-transformer/packages/format-switcher/tests/Converter/ConfigFormatConverter/FixtureYamlToPhp/maker-bundle/services_imports.yaml
@@ -6,5 +6,5 @@ imports:
 namespace Symfony\Component\DependencyInjection\Loader\Configurator;
 
 return static function (ContainerConfigurator $containerConfigurator): void {
-    $containerConfigurator->import(__DIR__ . '/services_simple.yaml', __DIR__ . '/annotations', true);
+    $containerConfigurator->import(__DIR__ . '/services_simple.yaml', 'annotations', true);
 };
